Moved term_colors.py to examples folder.
