+++ models/examples/term_colors.py
